Move into_sectors into new type
